+++ src/PageDetect.ts
@@ -2,11 +2,9 @@
 
 import * as select from 'select-dom';
 
-export const isGist = () =>
-  location.hostname.startsWith('gist.') || location.pathname.startsWith('gist/');
+export const isGist = () => location.hostname.startsWith('gist.') || location.pathname.startsWith('gist/');
 
-export const isDashboard = () =>
-  location.pathname === '/' || /^(\/orgs\/[^/]+)?\/dashboard/.test(location.pathname);
+export const isDashboard = () => location.pathname === '/' || /^(\/orgs\/[^/]+)?\/dashboard/.test(location.pathname);
 
 export const isTrending = () => location.pathname.startsWith('/trending');
 
@@ -40,11 +38,9 @@ export const isPR = () => isRepo() && /^\/pull\/\d+/.test(getRepoPath());
 
 export const isPRFiles = () => isRepo() && /^\/pull\/\d+\/files/.test(getRepoPath());
 
-export const isPRCommit = () =>
-  isRepo() && /^\/pull\/\d+\/commits\/[0-9a-f]{5,40}/.test(getRepoPath());
+export const isPRCommit = () => isRepo() && /^\/pull\/\d+\/commits\/[0-9a-f]{5,40}/.test(getRepoPath());
 
-export const isHistoryForFile = () =>
-  isRepo() && /^\/commits\/[0-9a-f]{5,40}\/.+/.test(getRepoPath());
+export const isHistoryForFile = () => isRepo() && /^\/commits\/[0-9a-f]{5,40}\/.+/.test(getRepoPath());
 
 export const isMilestoneList = () => isRepo() && /^\/milestones\/?$/.test(getRepoPath());
 
@@ -58,8 +54,7 @@ export const isCommitList = () => isRepo() && /^\/commits\//.test(getRepoPath())
 
 export const isSingleCommit = () => isRepo() && /^\/commit\/[0-9a-f]{5,40}/.test(getRepoPath());
 
-export const isCommit = () =>
-  isSingleCommit() || isPRCommit() || (isPRFiles() && select.exists('.full-commit'));
+export const isCommit = () => isSingleCommit() || isPRCommit() || (isPRFiles() && select.exists('.full-commit'));
 
 export const isCompare = () => isRepo() && /^\/compare/.test(getRepoPath());
 
@@ -69,11 +64,7 @@ export const hasCode = () => isRepo() && select.exists('.highlight');
 
 export const hasDiff = () =>
   isRepo() &&
-  (isSingleCommit() ||
-    isPRCommit() ||
-    isPRFiles() ||
-    isCompare() ||
-    (isPR() && select.exists('.diff-table')));
+  (isSingleCommit() || isPRCommit() || isPRFiles() || isCompare() || (isPR() && select.exists('.diff-table')));
 
 export const isReleases = () => isRepo() && /^\/(releases|tags)/.test(getRepoPath());